+++ Back/Java/firstStep/firstStep/src/main/java/com/matthewoks/firstStep/Models/Meet.java
@@ -1,22 +1,22 @@
 package com.matthewoks.firstStep.Models;
 
 public abstract class Meet {
-    private int Id;
+    private Long Id;
     private String color;
     private String meetName;
 public Meet(){}
 
-    public Meet(int id,String color, String meetName) {
+    public Meet(Long id,String color, String meetName) {
     super();
         this.Id=id;
         this.meetName = meetName;
         this.color = color;
     }
-    public int getId() {
+    public Long getId() {
         return Id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         Id = id;
     }
     public String getColor() {
